More Backend Changes
Added a section of sale validation to meet project requirements.

Defined a "get_current_holdings" function to calculate the sum of all transactions for either a bond or stock for an investor.

Updated the create_stock_transaction and create_bondtraction endpoints to see if the sale quantity exceeds current holdings. It should return an error if the rule is violated.